+++ demos/client-bundle-example/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prosopo/client-bundle-example
 
+## 2.9.1
+### Patch Changes
+
+- 86c22b8: structured logging
+
+## 2.9.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+## 2.8.3
+### Patch Changes
+
+- cf26d7e: Prevents the translation key (WIDGET.I_AM_HUMAN) to be shown until the translation is loaded
+
 ## 2.8.2
 
 ### Patch Changes