+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @prosopo/client-example-server
 
+## 2.7.6
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+  - @prosopo/dotenv@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/server@2.9.2
+
+## 2.7.5
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/server@2.9.1
+
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [d5f2e95]
+  - @prosopo/server@2.9.0
+
+## 2.7.3
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/common@2.7.2
+  - @prosopo/dotenv@2.6.4
+  - @prosopo/server@2.8.1
+  - @prosopo/types@2.9.1
+  - @prosopo/util@2.6.1
+
+## 2.7.2
+### Patch Changes
+
+- Updated dependencies [d6de900]
+  - @prosopo/server@2.8.0
+
+## 2.7.1
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+  - @prosopo/common@2.7.1
+  - @prosopo/server@2.7.1
+  - @prosopo/dotenv@2.6.3
+
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+  - @prosopo/server@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/dotenv@2.6.2
+
+## 2.6.4
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/dotenv@2.6.1
+  - @prosopo/server@2.6.4
+  - @prosopo/types@2.7.1
+
 ## 2.6.3
 
 ### Patch Changes