+++ packages/procaptcha-frictionless/CHANGELOG.md
@@ -1,5 +1,106 @@
 # @prosopo/procaptcha-frictionless
 
+## 2.6.19
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/detector@3.0.0
+  - @prosopo/locale@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/procaptcha-pow@2.7.5
+  - @prosopo/procaptcha-react@2.6.19
+
+## 2.6.18
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/procaptcha-pow@2.7.4
+  - @prosopo/procaptcha-react@2.6.18
+
+## 2.6.17
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/procaptcha-react@2.6.17
+  - @prosopo/procaptcha-pow@2.7.3
+  - @prosopo/types@2.9.1
+
+## 2.6.16
+### Patch Changes
+
+  - @prosopo/procaptcha-pow@2.7.2
+  - @prosopo/procaptcha-react@2.6.16
+
+## 2.6.15
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+  - @prosopo/procaptcha-pow@2.7.1
+  - @prosopo/procaptcha-react@2.6.15
+
+## 2.6.14
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/procaptcha-pow@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/procaptcha-react@2.6.14
+
+## 2.6.13
+### Patch Changes
+
+- cf26d7e: Prevents the translation key (WIDGET.I_AM_HUMAN) to be shown until the translation is loaded
+- Updated dependencies [cf26d7e]
+  - @prosopo/procaptcha-pow@2.6.13
+  - @prosopo/procaptcha-react@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- ea38a1c: lint
+- Updated dependencies [ea38a1c]
+  - @prosopo/procaptcha-react@2.6.12
+  - @prosopo/procaptcha-pow@2.6.12
+
+## 2.6.11
+
+### Patch Changes
+
+- b2ae723: lint
+- Updated dependencies [b2ae723]
+  - @prosopo/procaptcha-react@2.6.11
+  - @prosopo/procaptcha-pow@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- d17c67f: lint
+- Updated dependencies [d17c67f]
+  - @prosopo/procaptcha-react@2.6.10
+  - @prosopo/procaptcha-pow@2.6.10
+
+## 2.6.9
+
+### Patch Changes
+
+- 0d194f2: lint
+- Updated dependencies [0d194f2]
+  - @prosopo/procaptcha-react@2.6.9
+  - @prosopo/procaptcha-pow@2.6.9
+
+## 2.6.8
+
+### Patch Changes
+
+- @prosopo/procaptcha-pow@2.6.8
+- @prosopo/procaptcha-react@2.6.8
+- @prosopo/types@2.7.1
+
 ## 2.6.7
 
 ### Patch Changes