+++ packages/procaptcha/CHANGELOG.md
@@ -1,5 +1,125 @@
 # @prosopo/procaptcha
 
+## 2.7.5
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/datasets@3.0.0
+  - @prosopo/common@3.0.0
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+  - @prosopo/api@3.0.0
+  - @prosopo/account@2.7.4
+  - @prosopo/load-balancer@2.6.9
+  - @prosopo/procaptcha-common@2.7.4
+
+## 2.7.4
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+  - @prosopo/account@2.7.3
+  - @prosopo/api@2.7.2
+  - @prosopo/datasets@2.7.3
+  - @prosopo/load-balancer@2.6.8
+  - @prosopo/procaptcha-common@2.7.3
+
+## 2.7.3
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/procaptcha-common@2.7.2
+  - @prosopo/load-balancer@2.6.7
+  - @prosopo/util-crypto@13.5.1
+  - @prosopo/datasets@2.7.2
+  - @prosopo/account@2.7.2
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+  - @prosopo/api@2.7.1
+  - @prosopo/util@2.6.1
+
+## 2.7.2
+### Patch Changes
+
+- Updated dependencies [d6de900]
+  - @prosopo/api@2.7.0
+
+## 2.7.1
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/util-crypto@13.5.0
+  - @prosopo/types@2.9.0
+  - @prosopo/account@2.7.1
+  - @prosopo/common@2.7.1
+  - @prosopo/datasets@2.7.1
+  - @prosopo/api@2.6.6
+  - @prosopo/load-balancer@2.6.6
+  - @prosopo/procaptcha-common@2.7.1
+
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/procaptcha-common@2.7.0
+  - @prosopo/util-crypto@13.4.0
+  - @prosopo/datasets@2.7.0
+  - @prosopo/account@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/load-balancer@2.6.5
+  - @prosopo/api@2.6.5
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated dependencies [ea38a1c]
+  - @prosopo/procaptcha-common@2.6.11
+  - @prosopo/datasets@2.6.12
+
+## 2.6.11
+
+### Patch Changes
+
+- Updated dependencies [b2ae723]
+  - @prosopo/procaptcha-common@2.6.10
+  - @prosopo/datasets@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Updated dependencies [d17c67f]
+  - @prosopo/procaptcha-common@2.6.9
+  - @prosopo/datasets@2.6.10
+
+## 2.6.9
+
+### Patch Changes
+
+- Updated dependencies [0d194f2]
+  - @prosopo/procaptcha-common@2.6.8
+  - @prosopo/datasets@2.6.9
+
+## 2.6.8
+
+### Patch Changes
+
+- Updated dependencies [04cc7ee]
+  - @prosopo/common@2.6.1
+  - @prosopo/account@2.6.4
+  - @prosopo/datasets@2.6.8
+  - @prosopo/load-balancer@2.6.4
+  - @prosopo/procaptcha-common@2.6.7
+  - @prosopo/types@2.7.1
+  - @prosopo/api@2.6.4
+
 ## 2.6.7
 
 ### Patch Changes